StreamHPC 2024-01-16 (#365)
* feat(bench.cu): add partition to benchmarks

* ci: use build instead rocm-build tag

This allows the build job to be performed by any runner configured
for building, instead of the ROCm-specialized builder. As the
target architectures are specified ahead of time, the GPU is not
needed during the build process, and may be performed by any builder.

* perf(partition.h): use 'transform_iterator' instead of 'transform'

* remove some implicit conversions for -Wconversion

* test(bench.cu): improve benchmark accuracy for partition with floating points

* test(bench.cu): use 'std::back_inserter' instead of pre-allocating input-sized vectors

* test(bench.cu): remove cuda synchronization from partition benchmarks

Partition on CUDA is synchronous already, so device synchronization is unneeded.

* docs(bench.cu): add missing copyright header

* fix(type_traits.h): use 'rocprim::invoke_result' instead of 'rocprim::detail::invoke_result'

* docs(changelog.md): added internal usuage of 'rocprim::invoke_result' to the changelog

* Updated changelog style

* Specifying SPHINX_DIR in CI

* build: Updated version number

* Updated ROCm version number in changelog

---------


[ROCm/rocThrust commit: 1664895575e370345645d165bc63b09ba5e4fce4]

+++ thrust/system/hip/detail/internal/copy_cross_system.h
@@ -97,7 +97,7 @@ namespace __copy
                              derived_cast(sys2),
                              reinterpret_cast<InputTy*>(thrust::raw_pointer_cast(&*result)),
                              reinterpret_cast<InputTy const*>(thrust::raw_pointer_cast(&*begin)),
-                             n);
+                             static_cast<size_t>(n));
                      } return result + n;));
     }
 
@@ -283,7 +283,7 @@ cross_system_copy_n(thrust::hip_rocprim::execution_policy<D>& device_s,
                       InputIterator                  end,
                       OutputIterator                 result)
     {
-        return cross_system_copy_n(systems, begin, thrust::distance(begin, end), result);
+        return cross_system_copy_n(systems, begin, static_cast<size_t>(thrust::distance(begin, end)), result);
     }
 } // namespace __copy
 } // namespace hip_rocprim
